fix(MesosStateStore): adjust stream data handling
Use `Promise.resolve().then()` to schedule a micro tasks and asynchronously trigger `onStreamData` which is used to emit the `MESOS_STATE_CHANGE`. This change will mitigate issues in which errors in view components lead to stale data as the error propergation broke the stream.

N.B. `EventEmitter` is synchronous meaning that every delay in an event handler effects the `emit` execution time and errors propagate up to the `emit` call which sometimes leads to undesireable behavior.

Closes DCOS-21784